Fix imports.
Moved everything back into TotalConnectClient because I can't figure out
how to use separate files.
