+++ total_connect_client/__init__.py
@@ -1,1 +1,2 @@
 """Initialize."""
+
